some code fixes
