+++ news-app/src/main/java/com/seminav/newsapp/controllers/PingController.java
@@ -1,6 +1,6 @@
 package com.seminav.newsapp.controllers;
 
-import com.seminav.newsapp.external.services.AuthService;
+import com.seminav.newsapp.external.services.ExternalUserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 public class PingController {
-    private final AuthService authService;
+    private final ExternalUserService externalUserService;
 
     @GetMapping("/ping")
     public String ping() {
@@ -21,6 +21,6 @@ public class PingController {
             @RequestHeader("Authorization") String authorizationHeader
 
     ) {
-        return authService.getUserRoleFromUserAndValidateJWT(authorizationHeader);
+        return externalUserService.getUserRoleFromUserAndValidateJWT(authorizationHeader);
     }
 }
