--- news-app/src/main/java/com/seminav/newsapp/external/services/AuthService.java
@@ -1,5 +1,0 @@
-package com.seminav.newsapp.external.services;
-
-public interface AuthService {
-    String getUserRoleFromUserAndValidateJWT(String authHeader);
-}
