+++ news-app/src/main/java/com/seminav/newsapp/external/services/ExternalUserService.java
@@ -1,0 +1,9 @@
+package com.seminav.newsapp.external.services;
+
+import java.util.List;
+
+public interface ExternalUserService {
+    String getUserRoleFromUserAndValidateJWT(String authHeader);
+    List<String> getUserSavedNews(String authHeader);
+
+}
